refactor: Move all slashcmd files (#53)
