+++ src/commands/slashcommands/create-challenge.ts
@@ -1,16 +1,16 @@
 import { SlashCommandBuilder } from '@discordjs/builders';
 import { CommandInteractionOption, GuildMember, PermissionsBitField } from 'discord.js';
-import { addChallengeToTournament, getDifficultyByEmoji, getTournamentByName } from '../backend/queries/tournamentQueries.js';
-import { ChallengeDocument } from '../types/customDocument.js';
-import { ChallengeModel } from '../backend/schemas/challenge.js';
-import { OptionValidationError, OptionValidationErrorStatus } from '../types/customError.js';
-import { getCurrentTournament } from '../backend/queries/guildSettingsQueries.js';
-import { LimitedCommandInteraction } from '../types/limitedCommandInteraction.js';
-import { OptionValidationErrorOutcome, Outcome, OutcomeStatus, OutcomeWithDuoBody, SlashCommandDescribedOutcome } from '../types/outcome.js';
-import { ValueOf } from '../types/typelogic.js';
-import { Constraint, validateConstraints } from './slashcommands/architecture/validation.js';
-import { getJudgeByGuildIdAndMemberId } from '../backend/queries/profileQueries.js';
-import { SimpleRendezvousSlashCommand } from './slashcommands/architecture/rendezvousCommand.js';
+import { addChallengeToTournament, getDifficultyByEmoji, getTournamentByName } from '../../backend/queries/tournamentQueries.js';
+import { ChallengeDocument } from '../../types/customDocument.js';
+import { ChallengeModel } from '../../backend/schemas/challenge.js';
+import { OptionValidationError, OptionValidationErrorStatus } from '../../types/customError.js';
+import { getCurrentTournament } from '../../backend/queries/guildSettingsQueries.js';
+import { LimitedCommandInteraction } from '../../types/limitedCommandInteraction.js';
+import { OptionValidationErrorOutcome, Outcome, OutcomeStatus, OutcomeWithDuoBody, SlashCommandDescribedOutcome } from '../../types/outcome.js';
+import { ValueOf } from '../../types/typelogic.js';
+import { Constraint, validateConstraints } from './architecture/validation.js';
+import { getJudgeByGuildIdAndMemberId } from '../../backend/queries/profileQueries.js';
+import { SimpleRendezvousSlashCommand } from './architecture/rendezvousCommand.js';
 
 /**
  * Alias for the first generic type of the command.
